created leadership component
